Added support for RegExp in allowOrigins (#1)
* Added support for RegExp in allowOrigins

If setting 'allowOriginsRegExp' to true in the config it is possible to
specify RegExp rules in the 'allowOrign' array instead of urls.

* Review: RegExp no longer catches the match

* Added method for collecting CORS headers on Request object

Adding CORS headers to the HTTP response is the last thing we do in a
long chain of commands. If one fails (throws an exception) we risk not having the CORS headers
added. This ensures we have access to the relevant headers even if an error
should occure.

* Removed call to corsHeaders() in handleRequest()

The call was only there to please the test, so I moved it to the test
code...

* Cleanup outcommented code